requirements.txt
streamlit
reportlab
PyMuPDF
regex
psycopg2-binary
pandas
python-dotenv
langchain-core
langchain-community
langchain-ollama
sympy
numpy
plotly
pytest
sentence-transformers
